Move the list processing methods to a library module
