+++ test/integration/delete_from_list/inspec/delete_from_list_spec.rb
@@ -37,6 +37,14 @@ describe file('/tmp/emptyfile') do
   its(:size) { should eq 0 }
 end
 
+describe file('/tmp/ends_with_test_last_entry') do
+  its(:content) { should match(%r{GRUB_CMDLINE_LINUX="rd.lvm.lv=centos/root rd.lvm quiet"}) }
+end
+
+describe file('/tmp/ends_with_test_middle_entry') do
+  its(:content) { should match(%r{GRUB_CMDLINE_LINUX="rd.lvm.lv=centos/root quiet elevator=noop"}) }
+end
+
 describe file('/tmp/chef_resource_status') do
   its(:content) { should match(/missing_file fail.*n#{eol}/) }
   its(:content) { should match(/missing_file\]\s*n#{eol}/) }
